- intital commit with functioning yet not proved converging RL storage startegies
