+++ examples/inputs/example_02e/demand_units.csv
@@ -1,0 +1,2 @@
+name,technology,bidding_EOM,max_power,min_power,unit_operator
+demand_EOM,inflex_demand,naive_eom,1000000,0,eom_de
