+++ examples/inputs/example_02e/fuel_prices_df.csv
@@ -1,0 +1,2 @@
+fuel,uranium,lignite,hard coal,natural gas,oil,biomass,co2
+price,0.9,1.8,8.5,26,22,21,25
