+++ examples/inputs/example_02e/fuel_prices_df.csv.license
@@ -1,0 +1,3 @@
+SPDX-FileCopyrightText: ASSUME Developers
+
+SPDX-License-Identifier: AGPL-3.0-or-later
