+++ examples/inputs/example_02e/storage_units.csv
@@ -1,0 +1,2 @@
+name,technology,bidding_EOM,max_power_charge,max_power_discharge,efficiency_charge,efficiency_discharge,min_volume,max_volume,additional_cost_charge,additional_cost_discharge,natural_inflow,unit_operator
+Storage 1,PSPP,storage_learning,1000.0,992.0,0.86,0.9,1.0,6076.0,0.28,0.28,0.0,Operator 1
